refactor: extract shared chain/token fetch helpers from internal routes
